Time: 27 ms (67.34%), Space: 19.4 MB (49.35%) - LeetHub

+++ 36-valid-sudoku/36-valid-sudoku.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     bool isValidSudoku(vector<vector<char>>& board) {
-        unordered_map<int, bool> row;
-        unordered_map<int, bool> column;
-        unordered_map<int, bool> box;
+        unordered_map<short int, bool> row;
+        unordered_map<short int, bool> column;
+        unordered_map<short int, bool> box;
         for(int i = 0; i < 9; i++){
             for(int j = 0; j < 9; j++){
                 if(board[i][j] == '.'){
